reorganized and removed unecessary example code
