+++ src/resources/fxml/Main.fxml
@@ -31,7 +31,6 @@
 
 <StackPane fx:id="root" xmlns:fx="http://javafx.com/fxml/1"
 	xmlns="http://javafx.com/javafx/2.2">
-	<!-- Header -->
 	<BorderPane>
 
 		<top>
@@ -55,17 +54,8 @@
 			</VBox>
 		</top>
 
-		<!-- Content Area -->
 		<center>
-			<JFXDrawer fx:id="drawer" defaultDrawerSize="250" direction="LEFT">
-				<styleClass>
-					<String fx:value="body" />
-				</styleClass>
-				<content>
-					<StackPane fx:id="content" >
-					</StackPane>
-				</content>
-			</JFXDrawer>
+      <StackPane fx:id="content" />
 		</center>
 	</BorderPane>
